requirements.txt
aspose.words
aspose.slides